fix: set max Metaspace size to 512MB
Keep the original value that was present in `scalingo.json`.
Make sure `bin/start` is aligned with this value.

This value seems rather conservative. Anyway, the only way to size the Metaspace
seems to consist in running the app for a few days, get the Garbage Collector
to log what it does and analyze these logs to adjust the value. So we
can't put a (default) optimized value here.